+++ storage/src/merkle/mmr/hasher.rs
